+++ client/src/app/angular-material.module.ts
@@ -12,6 +12,9 @@ import { MatListModule } from '@angular/material/list';
 import { MatGridListModule } from '@angular/material/grid-list';
 import { MatSelectModule } from '@angular/material/select';
 import { MatButtonModule } from '@angular/material/button';
+import { MatTabsModule } from '@angular/material/tabs';
+import { MatMenuModule } from '@angular/material/menu';
+import { MatDialogModule } from '@angular/material/dialog';
 
 const materialModules = [
   CdkTreeModule,
@@ -25,7 +28,10 @@ const materialModules = [
   MatCheckboxModule,
   MatGridListModule,
   MatSelectModule,
-  MatButtonModule
+  MatButtonModule,
+  MatTabsModule,
+  MatMenuModule,
+  MatDialogModule
 ];
 @NgModule({
   imports: [CommonModule, ...materialModules],