+++ client/src/app/app.component.spec.ts
@@ -1,12 +1,25 @@
 import { TestBed } from '@angular/core/testing';
+import { AngularFireModule } from '@angular/fire/compat';
+import { AngularFirestoreModule } from '@angular/fire/compat/firestore';
+import { MatButtonModule } from '@angular/material/button';
+import { MatRippleModule } from '@angular/material/core';
+import { MatTabNavPanel } from '@angular/material/tabs';
 import { RouterTestingModule } from '@angular/router/testing';
+import { environment } from 'src/env/environments';
 import { AppComponent } from './app.component';
+import { LayoutComponent } from './components/layout/layout.component';
 
 describe('AppComponent', () => {
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports: [RouterTestingModule],
-      declarations: [AppComponent]
+      imports: [
+        RouterTestingModule,
+        AngularFireModule.initializeApp(environment.firebase),
+        AngularFirestoreModule,
+        MatButtonModule,
+        MatRippleModule
+      ],
+      declarations: [AppComponent, LayoutComponent, MatTabNavPanel]
     }).compileComponents();
   });
 
@@ -16,9 +29,9 @@ describe('AppComponent', () => {
     expect(app).toBeTruthy();
   });
 
-  it(`should have as title 'to-do-list'`, () => {
+  it(`should have as title 'My-To-Dos'`, () => {
     const fixture = TestBed.createComponent(AppComponent);
     const app = fixture.componentInstance;
-    expect(app.title).toEqual('to-do-list');
+    expect(app.title).toEqual('My-To-Dos');
   });
 });