+++ client/src/app/components/layout/layout.component.ts
@@ -1,0 +1,13 @@
+import { Component } from '@angular/core';
+import { AuthService } from 'src/app/services/auth.service';
+
+@Component({
+  selector: 'app-layout',
+  templateUrl: './layout.component.html',
+  styleUrls: ['./layout.component.css']
+})
+export class LayoutComponent {
+  constructor(public authService: AuthService) {}
+
+  ngOnInit(): void {}
+}