+++ client/src/app/services/usage.service.spec.ts
@@ -1,0 +1,37 @@
+import { TestBed } from '@angular/core/testing';
+import {
+  HttpClientTestingModule,
+  HttpTestingController
+} from '@angular/common/http/testing';
+import { AngularFireModule } from '@angular/fire/compat';
+import { AngularFirestoreModule } from '@angular/fire/compat/firestore';
+import { environment } from 'src/env/environments';
+import { RouterTestingModule } from '@angular/router/testing';
+import { HomeComponent } from '../components/home/home.component';
+
+import { UsageService } from './usage.service';
+
+describe('UsageService', () => {
+  let service: UsageService;
+  let httpTestingController: HttpTestingController;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      imports: [
+        HttpClientTestingModule,
+        AngularFireModule.initializeApp(environment.firebase),
+        AngularFirestoreModule,
+        RouterTestingModule.withRoutes([
+          { path: 'home', component: HomeComponent }
+        ])
+      ]
+    });
+
+    service = TestBed.inject(UsageService);
+    httpTestingController = TestBed.inject(HttpTestingController);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});